ajout des controller parrain, utilisateur, association et authentification des associations

+++ src/main/java/com/djiguiya/djiguiya/repository/EnfantRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface EnfantRepository extends JpaRepository<Enfant, Integer> {
+public interface EnfantRepository extends JpaRepository<Enfant, Long> {
 }
