+++ src/main/java/com/djiguiya/djiguiya/repository/ParentRepository.java
@@ -4,5 +4,5 @@ import com.djiguiya.djiguiya.entity.Parent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ParentRepository extends JpaRepository<Parent, Integer> {
+public interface ParentRepository extends JpaRepository<Parent, Long> {
 }
