+++ src/main/java/com/djiguiya/djiguiya/repository/ParrainRepository.java
@@ -4,5 +4,5 @@ import com.djiguiya.djiguiya.entity.Parrain;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ParrainRepository extends JpaRepository<Parrain, Integer> {
+public interface ParrainRepository extends JpaRepository<Parrain, Long> {
 }
